rename /api/* folders
