+++ lib/api/boot/index.js
@@ -12,19 +12,19 @@ var app = module.exports = express()
  * Local signin routes
  */
 
-app.use('/signin', require('lib/api/signin-api'))
+app.use('/signin', require('lib/api/signin'))
 
 /*
  * Local signup routes
  */
 
-if (!config.signupUrl) app.use('/signup', require('lib/api/signup-api'))
+if (!config.signupUrl) app.use('/signup', require('lib/api/signup'))
 
 /*
  * Forgot password routes
  */
 
-app.use('/forgot', require('lib/api/forgot-api'))
+app.use('/forgot', require('lib/api/forgot'))
 
 /**
  * Root API Service
@@ -48,13 +48,13 @@ app.all('*', visibility)
  * Account routes
  */
 
-app.use('/settings', require('lib/api/settings-api'))
+app.use('/settings', require('lib/api/settings'))
 
 /*
  * Stats routes
  */
 
-app.use('/stats', require('lib/api/stats-api'))
+app.use('/stats', require('lib/api/stats'))
 
 /*
  * RSS routes
@@ -72,28 +72,28 @@ app.use('/api', require('lib/tag'))
  * Topic API Service
  */
 
-app.use('/api', require('lib/api/topic-api'))
+app.use('/api', require('lib/api/topic'))
 
 /**
  * Whitelist API Service
  */
 
-app.use('/api', require('lib/api/whitelist-api'))
+app.use('/api', require('lib/api/whitelist'))
 
 /**
  * Comment API Service
  */
 
-app.use('/api', require('lib/api/comment-api'))
+app.use('/api', require('lib/api/comment'))
 
 /**
  * Forums API Service
  */
 
-app.use('/api', require('lib/api/forum-api'))
+app.use('/api', require('lib/api/forum'))
 
 /**
  * Notifications API Service
  */
 
-app.use('/api', require('lib/api/notification-api'))
+app.use('/api', require('lib/api/notification'))
