+++ lib/api/signin/index.js
