+++ lib/api/signup/index.js
