+++ lib/api/signup/lib/signup.js
